qc/modules/strandedness: consolidate pipeline tests for 'strandedness' module.
